refactor(services): Extract TeamMemberAccessService from TeamService
- Created TeamMemberAccessService (182 lines) as atomic service for team member access verification
- Extracted verifyTeamAccess, getUserActiveTeamCount, getTeamMemberCount from TeamService
- Added helper methods: isTeamMember, getUserTeamRole, isTeamAdmin
- Updated TeamService to use teamMemberAccessService instead of private methods
- All access verification and counting logic now centralized
- Zero breaking changes - all public interfaces preserved
- Code quality: Reduced TeamService from 1466 to 1371 lines (95 lines removed)
- All 69 test suites passing (1136/1169 tests, 33 todo)

Part of TeamService refactoring (Step 2/5)
